Mostly changes to 2H weaps
-Added some hooks for 2H weaps for future customization.

Bugs:
-DrawEffects() for two handers works really bad in multiplayer. Not sure about others tho.

+++ Projectiles/Misc/TwoHandedWeaponHitbox.cs
@@ -7,14 +7,21 @@ using System.Text;
 using System.Threading.Tasks;
 using Terraria;
 using Terraria.ID;
+using TLoZ.Items.Weapons;
 using TLoZ.Items.Weapons.NoGroup;
 
 namespace TLoZ.Projectiles.Misc
 {
     public class TwoHandedWeaponHitbox : TLoZProjectile
     {
+        private TwoHandedWeapon Weapon => Owner.HeldItem.modItem as TwoHandedWeapon;
+        public override void SetStaticDefaults()
+        {
+            DisplayName.SetDefault("");
+        }
         public override void SetDefaults()
         {
+            projectile.melee = true;
             projectile.tileCollide = false;
             projectile.width = 100;
             projectile.height = 120;
@@ -24,27 +31,26 @@ namespace TLoZ.Projectiles.Misc
         public override string Texture => "TLoZ/Textures/Misc/EmptyPixel";
         public override void AI()
         {
-            projectile.timeLeft = 2;
-            if (!TLoZPlayer.isSwinging)
-                projectile.Kill();
             Texture2D texture = Main.itemTexture[Owner.HeldItem.type];
-            projectile.width = (int)(texture.Width * 2f);
-            projectile.Center = Owner.Center + new Vector2(texture.Width * .5f * Owner.direction, -20);
 
-            if(Owner.HeldItem.type == mod.ItemType<WoodcutterAxe>())
-            {
-                for(int i = 0; i < projectile.height; i++)
-                {
-                    for(int j = 0; j < projectile.width; j++)
-                    {
-                        Tile tile = Main.tile[(int)(projectile.position.X + j) / 16, (int)(projectile.position.Y + i) / 16];
-                        if (tile.type == TileID.Trees || tile.type == TileID.PalmTree)
-                        {
-                            WorldGen.KillTile((int)(projectile.position.X + j) / 16, (int)(projectile.position.Y + i) / 16, false, false);
-                        }
-                    }
-                }
-            }
+            projectile.timeLeft = 2;
+
+            if (!TLoZPlayer.isSwinging && !TLoZPlayer.downwardsSlash)
+                projectile.Kill();
+
+            projectile.width = (int)((texture.Width * 0.5f + texture.Height * 0.5f) * 2f);
+
+            projectile.Center = Owner.Center + new Vector2(texture.Width * .75f * Owner.direction, TLoZPlayer.downwardsSlash ? 0 : -20);
+
+            if (TLoZPlayer.downwardsSlash)
+                projectile.height = 140;
+
+            Weapon?.HitboxEffects(projectile, Owner);
+        }
+
+        public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
+        {
+            Weapon?.OnHitEffects(target, Owner);
         }
     }
 }
